Fix compilation error on linux

+++ src/lib/Database/DbQueryBinding.cpp
@@ -187,7 +187,7 @@ bool DbQueryBinding::getString(DbConnection* connection, int columnIndex, std::s
 		outString->clear();
 		return true;
 	} else if(dataSize < 0) {
-		logStatic(LL_Warning, DbQueryBinding::getStaticClassName(), "getString: dataSize is negative: %d\n", dataSize);
+		logStatic(LL_Warning, DbQueryBinding::getStaticClassName(), "getString: dataSize is negative: %lld\n", (int64_t)dataSize);
 		return false;
 	}
 
@@ -213,7 +213,7 @@ bool DbQueryBinding::getString(DbConnection* connection, int columnIndex, std::s
 		outString->clear();
 		return true;
 	} else if(isDataNull < 0) {
-		logStatic(LL_Warning, DbQueryBinding::getStaticClassName(), "getString: isDataNull is negative: %d\n", isDataNull);
+		logStatic(LL_Warning, DbQueryBinding::getStaticClassName(), "getString: isDataNull is negative: %lld\n", (int64_t)isDataNull);
 		return false;
 	} else if(ret == SQL_SUCCESS) {
 		bytesRead += isDataNull;
